Support for Action in DriverTestWidget. Also docstring are shown in tooltips
Actions are shown in a combobox at the bottom of DriverTest with a Run button
next to in. When run is pressed, a dialog opens asking for the value of each
argument.

In DriverTestWidget, the docstring for each feat is shown in the tooltip of
the corresponding label. When querying for function arguments, the tooltip shows
the part of the docstring that corresponds to that parameter.