Fixed UI support for Actions
To request argument values for a function via de UI, the callable
is introspected for argument names, default values and docstring.
In decorated functions, the __wrapped__ argument contains a reference
to the actual function.

Actions are called by returning Action.call curryied with the driver
instance when Action.__get__ is called.
This patch adds a __wrapped__ attribute to the returned partial function
pointing to the function wrapped by Action (Note that this is Action.func,
and not Action.cal, the curried function!)

This might be much easier once that PEP-362 (function signature object)
is implemented.