Changed to a more efficient logging scheme
Until this patch, message formatting in logging was peformed
for every message, independently if it was going to be shown
or not. Now message format and arguments are passed to the
logger which will build only those messages that are going to
be displayed.

To fully support PEP3101 messages and as the python logging
module uses % string interpolation internally lantz.log.get_logger
function analogous to logging.getLogger is provided that will
patch the returned logger with a PEP3101 compatible.

Due to an bug in Python 3.2, this patch requires Python 3.2.1