Instance specific modifiers of Feat, DictFeat and Actions
A Feat/Action object is shared among all members of the class
precluding a direct way to change the modifiers (e.g. units)
of a feat in only one instance. Instead of creating a new
instance of Feat/Action for each driver instance, this patch
provides:

 1.- A way of getting instance aware proxy of Feat/Action
     >>> driver.feats[feat_name]
     >>> driver.actions[action_name]
 2.- A easy way of storing in the instance (not in the feat)
     instance specific feat modifiers
     >>> driver.feats[feat_name].units = new_units
 3.- Changes to Feat/DictFeat/Action to use the instance
     information if present falling back to the shared info

Instance information is only stored on demand, no copies are
made until the functionality is needed.