New Self class to provide referencing between Feats
It is quite common that certain Feat attributes (e.g. units) are
dependent on the value of another Feat. This patch allows to create
such relationships. The Self object can only be used within a Feat
definition and takes the name of the related feat an in parenthesis
the default value. For example:

   @Feat(units=Self.units('ms'))
   def frequency(self):
       # The code goes here

   @Feat()
   def units(self):
       # The code goes here