Added LIBRARY_PREFIX and preprocess arguments hook
LibraryDriver now accepts a new Class attribute (LIBRARY_PREFIX) that
is passed to Library upon initialization. When a function is looked at
Library it will check LIBRARY_PREFIX + function_name in addition to
function_name. This removes much of the common boilerplate naming due
to the lack of namespaces in C code

The code that preprocess the arguments before calling the function has
been moved to a separated function _preprocess_args to enable easy
overriding.