Preliminary support for NI-DAQmx
A System class provides access to general information about the
installed devices and globally defined channels and tasks. It contains
three read-only dictionaries (channels, tasks, devices) that can be
used not only to obtain their names but to obtain instances of wrapping
object.

Channels, Tasks and Devices also provide these type of dictionaries for
accessing the related properties. A common dictionary is used to protect
for creating two wrappers of the same underlying physical channel/device/task.

Specialized clasess are defined in channels.py and tasks.py. This is still
work in progress. Many of the calls need to be moved from the base classes
to the specialized and viceversa. Function calls and parameters types still
need to be adapted.