initialize_many and finalize_many, and GUI support
This commit provides helper function for initialization and finalization
of an iterable of drivers. It also provides a helper function to simplify
the process of creating responsive GUI for initialization.

The function support dependencies between drivers (i.e. some driver(s) must
be initialized before another) and concurrency (drivers can be initialized
together)